Allow for setting communication protocol.
It is now possible to configure the protocol to use for communication with
github. To this end, users can specify the protocol in the repository
configuration, e.g.,:

  github:
    - id: mavam/gitdub
      protocol: ssh # Use ssh instead of git

Gitdub still defaults to "git" as protocol. Other options are "http" and "ssh".
The latter makes it possible to use gitdub also for private repositories.

Closes #2.